dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Node-appendChild.html

<!DOCTYPE html>
<title>Node.appendChild</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src=about:blank></iframe>
<script>
function testHRE(node) {
  test(function() {
    assert_throws("HIERARCHY_REQUEST_ERR", function() { node.appendChild(document.createTextNode("fail")) })
    assert_throws(new TypeError(), function() { node.appendChild(null) })
  })
}
test(function() {
  var frameDoc = frames[0].document
  var s = frameDoc.createElement("a")
  assert_equals(s.ownerDocument, frameDoc)
  document.body.appendChild(s)
  assert_equals(s.ownerDocument, document)

  assert_throws("HIERARCHY_REQUEST_ERR", function() { document.body.appendChild(frameDoc) })

  s = frameDoc.createElement("b")
  frameDoc.body.appendChild(s)
  assert_equals(s.ownerDocument, frameDoc)
  document.body.appendChild(s)
  assert_equals(s.ownerDocument, document)

  assert_throws(new TypeError(), function() { document.body.appendChild(null) })

  assert_throws(new TypeError(), function() { document.body.appendChild({'a':'b'}) })

  testHRE(document.createTextNode("Foo"));
  testHRE(document.createComment("Foo"));
  testHRE(document.doctype);
})
</script>
